server/functions/slack/requirements.txt
python-dotenv==0.13.0
requests==2.23.0
slackclient==2.6.1
Flask==1.1.2